+++ backend/dataall/modules/mlstudio/services/mlstudio_service.py
@@ -9,7 +9,7 @@ from dataclasses import dataclass, field
 from typing import List, Dict
 
 from dataall.base.context import get_context
-from dataall.core.environment.env_permission_checker import has_group_permission
+from dataall.core.environment.decorators.env_permission_checker import has_group_permission
 from dataall.core.environment.services.environment_service import EnvironmentService
 from dataall.core.permissions.db.resource_policy_repositories import ResourcePolicy
 from dataall.core.permissions.permission_checker import has_resource_permission, has_tenant_permission
