+++ backend/dataall/modules/notebooks/services/notebook_service.py
@@ -10,7 +10,7 @@ from typing import List, Dict
 
 from dataall.base.context import get_context as context
 from dataall.core.environment.db.environment_models import Environment
-from dataall.core.environment.env_permission_checker import has_group_permission
+from dataall.core.environment.decorators.env_permission_checker import has_group_permission
 from dataall.core.environment.services.environment_service import EnvironmentService
 from dataall.core.permissions.db.resource_policy_repositories import ResourcePolicy
 from dataall.core.permissions.permission_checker import has_resource_permission, has_tenant_permission
